reafactor: rename inference module files
